+++ src/main/java/com/whattobake/api/Service/LikesService.java
@@ -25,7 +25,7 @@ public class LikesService {
         });
     }
 
-    public Flux<Recipe> getLikedRecipesPages(String pbUid, Integer page) {
+    public Flux<Recipe> getLikedRecipesPages(String pbUid, Long page) {
         return likeRepository.getRecipesPages(pbUid, page);
     }
